requirement.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
streamlit==0.64.0
h5py==2.10.0
Pillow==7.1.2
numpy==1.18.1
tensorflow==2.3.0